refactor(preset): update regexManagers: to customManagers:
