+++ lib/config/presets/internal/index.ts
@@ -1,5 +1,6 @@
 import type { Preset, PresetConfig } from '../types';
 import * as configPreset from './config';
+import * as customManagersPreset from './custom-managers';
 import * as defaultPreset from './default';
 import * as dockerPreset from './docker';
 import * as groupPreset from './group';
@@ -9,7 +10,6 @@ import * as monorepoPreset from './monorepo';
 import * as npm from './npm';
 import * as packagesPreset from './packages';
 import * as previewPreset from './preview';
-import * as regexManagersPreset from './regex-managers';
 import * as replacements from './replacements';
 import * as schedulePreset from './schedule';
 import * as securityPreset from './security';
@@ -19,6 +19,7 @@ import * as workaroundsPreset from './workarounds';
 
 export const groups: Record<string, Record<string, Preset>> = {
   config: configPreset.presets,
+  customManagers: customManagersPreset.presets,
   default: defaultPreset.presets,
   docker: dockerPreset.presets,
   group: groupPreset.presets,
@@ -28,7 +29,6 @@ export const groups: Record<string, Record<string, Preset>> = {
   npm: npm.presets,
   packages: packagesPreset.presets,
   preview: previewPreset.presets,
-  regexManagers: regexManagersPreset.presets,
   replacements: replacements.presets,
   schedule: schedulePreset.presets,
   security: securityPreset.presets,
